# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Decortee API",
    "description": "Interior Design API",
    "version": "v1"
  },
  "paths": {
...
    "/api/Profile": {
      "get": {
        "tags": [
          "Profile"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Profile"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Profile/password": {
      "put": {
        "tags": [
          "Profile"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Profile/upload-image": {
      "post": {
        "tags": [
          "Profile"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "image": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "image": {
                  "style": "form"
                }
              }
            },
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "image": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "image": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Profile/theme": {
      "put": {
        "tags": [
          "Profile"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateThemeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Profile/language": {
      "put": {
        "tags": [
          "Profile"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLanguageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "OldPassword": {
            "type": "string"
          },
          "NewPassword": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "UpdateLanguageDto": {
        "type": "object",
        "properties": {
          "Language": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "Name": {
            "type": "string"
          },
          "Email": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "UpdateThemeDto": {
        "type": "object",
        "properties": {
          "DarkMode": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}